requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
celery
redis
requests
django-cors-headers
python-dotenv
drf-yasg
ipython
